requirements_dev.txt
langchain==0.3.8
langchain-community==0.3.8
langchain_groq==0.2.1
fpdf2==2.8.1
python-dotenv==1.0.1
atproto==0.0.55
cloudinary==1.41.0
markdown2==2.5.1
python-docx==1.1.2